+++ changelog.md
@@ -1,3 +1,7 @@
+##### 1.0.2
+- Simplified the lights icon
+- Removed the option to reset the settings
+
 ##### 1.0.1
 - Moved "Player light switch" to the top (just below "name") in individual light item settings